+++ TestProjects/BakeHLOD/Assets/Tests/EditMode/EditMode.asmdef.meta
@@ -1,0 +1,7 @@
+fileFormatVersion: 2
+guid: 949bd4925ddb41c41bff19a4a9c3c597
+AssemblyDefinitionImporter:
+  externalObjects: {}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
