--- TestProjects/BakeHLOD/Assets/Tests/HLODEditModeTests.cs.meta
@@ -1,11 +1,0 @@
-fileFormatVersion: 2
-guid: ed2b7a3b96ccd1447ab7f49f4b89cae9
-MonoImporter:
-  externalObjects: {}
-  serializedVersion: 2
-  defaultReferences: []
-  executionOrder: 0
-  icon: {instanceID: 0}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
