--- TestProjects/BakeHLOD/Assets/Tests/Tests.asmdef.meta
@@ -1,7 +1,0 @@
-fileFormatVersion: 2
-guid: 9c3a20f5f3977e84b9748ef52c5a585c
-AssemblyDefinitionImporter:
-  externalObjects: {}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
